testcontainers/requirements.txt
# Enhanced TestContainers Workshop - Essential Dependencies
# ========================================================

# Core TestContainers
testcontainers>=4.0.0

# Database Drivers
sqlalchemy>=2.0.0
redis>=5.0.0
pymongo>=4.5.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-html>=3.2.0

# System & Platform
docker>=6.1.0
psutil>=5.9.0
requests>=2.31.0
colorama>=0.4.6
tabulate>=0.9.0

# Optional: Enhanced features
# Uncomment if you want additional features
# matplotlib>=3.7.0  # For performance graphs
# rich>=13.0.0        # For enhanced terminal output
# click>=8.1.0        # For CLI enhancements

flask
requests